requirements.txt
# 基础科学计算
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
tqdm>=4.62.0

# 深度学习框架
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.20.0
accelerate>=0.12.0

# 分子建模和化学信息学
rdkit-pypi>=2022.3.0
selfies>=2.1.0
pymatgen>=2022.7.0
ase>=3.22.0

# 数据处理和存储
pyarrow>=8.0.0
h5py>=3.7.0
sqlite3  # 内置模块

# 机器学习和优化
pytorch-lightning>=1.7.0
wandb>=0.13.0
optuna>=3.0.0

# 多尺度计算
lammps-python>=3.4.0  # LAMMPS Python 接口
custodian>=2022.5.0   # VASP 作业管理
fipy>=3.4.0          # 相场模拟

# 强化学习
gym>=0.21.0
stable-baselines3>=1.6.0

# 可视化和分析
plotly>=5.9.0
ipywidgets>=7.7.0
jupyter>=1.0.0

# 工具和实用程序
pyyaml>=6.0
click>=8.0.0
pathlib  # 内置模块
typing-extensions>=4.0.0
dataclasses  # Python 3.7+ 内置

# 可选：高性能计算
# mpi4py>=3.1.0  # 如果需要 MPI 支持
# numba>=0.56.0  # JIT 编译加速 